feat: cleanup directive classes
