+++ webex_assistant_skills_sdk/api/directives/assistant_event.py
@@ -1,0 +1,26 @@
+from ast import Dict
+from typing import Any
+
+from pydantic import BaseModel
+
+from webex_assistant_skills_sdk.shared.models import Directive
+
+
+class AssistantEventPayload(BaseModel):
+    name: str
+    payload: Dict[str, Any]
+
+
+class AssistantEvent(Directive):
+    name: str = 'assistant-event'
+    payload: AssistantEventPayload
+
+    def __init__(
+        self,
+        name: str,
+        payload: Dict[str, Any],
+    ) -> None:
+        super().__init__(payload=AssistantEventPayload(
+            name=name,
+            payload=payload,
+        ))
