+++ webex_assistant_skills_sdk/shared/models/__init__.py
@@ -1,8 +1,10 @@
+from webex_assistant_skills_sdk.shared.models.base_directive import Directive
 from webex_assistant_skills_sdk.shared.models.context import DeviceContext
 from webex_assistant_skills_sdk.shared.models.dialogue import Dialogue
 
 
 __all__ = [
     'DeviceContext',
-    'Dialogue'
+    'Dialogue',
+    'Directive',
 ]
