--- webex_assistant_skills_sdk/shared/models/directives/__init__.py
@@ -1,28 +1,0 @@
-from webex_assistant_skills_sdk.shared.models.directives.base_directive import Directive
-from webex_assistant_skills_sdk.shared.models.directives.asr_hint import ASRHint, ASRHintPayload
-from webex_assistant_skills_sdk.shared.models.directives.assistant_event import AssistantEvent, AssistantEventPayload
-from webex_assistant_skills_sdk.shared.models.directives.clear_web_view import ClearWebView
-from webex_assistant_skills_sdk.shared.models.directives.display_web_view import DisplayWebView, DisplayWebViewPayload
-from webex_assistant_skills_sdk.shared.models.directives.listen import Listen
-from webex_assistant_skills_sdk.shared.models.directives.reply import Reply, ReplyPayload
-from webex_assistant_skills_sdk.shared.models.directives.speak import Speak, SpeakPayload
-from webex_assistant_skills_sdk.shared.models.directives.ui_hint import UIHint, UIHintPayload
-
-
-__all__ = [
-    'Directive',
-    'ASRHint',
-    'ASRHintPayload',
-    'AssistantEvent',
-    'AssistantEventPayload',
-    'ClearWebView',
-    'DisplayWebView',
-    'DisplayWebViewPayload',
-    'Listen',
-    'Reply',
-    'ReplyPayload',
-    'Speak',
-    'SpeakPayload',
-    'UIHint',
-    'UIHintPayload'
-]
