--- webex_assistant_skills_sdk/shared/models/directives/clear_web_view.py
@@ -1,5 +1,0 @@
-from webex_assistant_skills_sdk.shared.models.directives.base_directive import Directive
-
-
-class ClearWebView(Directive):
-    name: str = 'clear-web-view'
