--- webex_assistant_skills_sdk/shared/models/directives/listen.py
@@ -1,5 +1,0 @@
-from webex_assistant_skills_sdk.shared.models.directives.base_directive import Directive
-
-
-class Listen(Directive):
-    name: str = 'listen'
